:ambulance: to ensure that the message is processed only once

+++ src/main/java/io/github/biezhi/wechat/api/model/Member.java
@@ -28,7 +28,7 @@ public class Member {
     private String remarkName;
 
     @SerializedName("AttrStatus")
-    private Integer attrStatus;
+    private Long attrStatus;
 
     @SerializedName("MemberStatus")
     private Integer memberStatus;
